fix: Secure recipe update and delete

+++ src/main/java/md/ctif/recipes_app/service/RecipeService.java
@@ -6,7 +6,11 @@ import md.ctif.recipes_app.entity.Recipe;
 import md.ctif.recipes_app.repository.RecipeRepository;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.ReactiveSecurityContextHolder;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
@@ -55,19 +59,47 @@ public class RecipeService {
         return recipeRepository.findById(id);
     }
 
-    public Mono<Void> deleteById(Long id) {
-        return recipeRepository.deleteById(id);
+    public Mono<ResponseEntity<String>> deleteById(Long id) {
+        return ReactiveSecurityContextHolder.getContext()
+                .map(SecurityContext::getAuthentication)
+                .map(auth -> {
+                    if (auth.getPrincipal() instanceof Jwt jwt) {
+                        return jwt.getClaimAsString("sub");
+                    }
+                    return auth.getName();
+                }).flatMap(sub -> recipeRepository.findById(id)
+                        .switchIfEmpty(Mono.error(new RuntimeException("Recipe not found with ID: " + id)))
+                        .flatMap(existingRecipe -> {
+                                    if (existingRecipe.getCreatedBy().equals(sub)) {
+                                        return recipeRepository.deleteById(id)
+                                                .then(Mono.just(ResponseEntity.ok("Recipe deleted with ID: " + id)));
+                                    }
+                                    return Mono.just(ResponseEntity.status(HttpStatus.FORBIDDEN).build());
+                                }
+                        )
+                );
     }
 
 
     public Mono<ResponseEntity<String>> update(Long id, RecipeDTO recipeDTO, String imagePath) {
-        return recipeRepository.findById(id)
-                .switchIfEmpty(Mono.error(new RuntimeException("Recipe not found with ID: " + id)))
-                .flatMap(existingRecipe -> {
-                    existingRecipe.update(recipeDTO.withImage(imagePath));
-                    return updateIngredientsTagsAndRecipe(recipeDTO, existingRecipe)
-                            .then(Mono.just(ResponseEntity.ok("Recipe updated")));
-                });
+        return ReactiveSecurityContextHolder.getContext()
+                .map(SecurityContext::getAuthentication)
+                .map(auth -> {
+                    if (auth.getPrincipal() instanceof Jwt jwt) {
+                        return jwt.getClaimAsString("sub");
+                    }
+                    return auth.getName();
+                }).flatMap(sub -> recipeRepository.findById(id)
+                        .switchIfEmpty(Mono.error(new RuntimeException("Recipe not found with ID: " + id)))
+                        .flatMap(existingRecipe -> {
+                            if (existingRecipe.getCreatedBy().equals(sub)) {
+                                existingRecipe.update(recipeDTO.withImage(imagePath));
+                                return updateIngredientsTagsAndRecipe(recipeDTO, existingRecipe)
+                                        .then(Mono.just(ResponseEntity.ok("Recipe updated")));
+                            } else {
+                                return Mono.just(ResponseEntity.status(HttpStatus.FORBIDDEN).build());
+                            }
+                        }));
     }
 
     private Mono<Void> updateIngredientsTagsAndRecipe(RecipeDTO recipeDTO, Recipe existingRecipe) {
